Update description of my guide
My guide was written for MEW, not MyCrypto, so the title should reflect that to prevent confusion.  I made the mistake of putting the guide on Steemit, so I can't edit it to reflect these changes. I do intend on remaking it in the future, and will do so on the MyCrypto platform.  I've been waiting on MEW v4 to do this, but I will wait until an equivalent update on MyCrypto happens.

- Tomshwom | AtLeastSignificant